backend/requirements.txt
fastapi
uvicorn
pandas
google-generativeai
python-dotenv
pypdf
python-docx
sqlalchemy
